feat: move config to its own "module"
