+++ packages/cli/src/modules/config/commands/print.ts
@@ -17,7 +17,7 @@
 import { OptionValues } from 'commander';
 import { stringify as stringifyYaml } from 'yaml';
 import { AppConfig, ConfigReader } from '@backstage/config';
-import { loadCliConfig } from '../../lib/config';
+import { loadCliConfig } from '../lib/config';
 import { ConfigSchema, ConfigVisibility } from '@backstage/config-loader';
 
 export default async (opts: OptionValues) => {
