+++ packages/cli/src/modules/config/commands/validate.ts
@@ -15,7 +15,7 @@
  */
 
 import { OptionValues } from 'commander';
-import { loadCliConfig } from '../../lib/config';
+import { loadCliConfig } from '../lib/config';
 
 export default async (opts: OptionValues) => {
   await loadCliConfig({
